Migration of mgmtsystem_system module from 8.0 to 9.0 module no more dependant from document_page
