+++ mgmtsystem/models/__init__.py
@@ -19,19 +19,5 @@
 #
 ##############################################################################
 
-from openerp import models, fields
 
-
-def own_company(self):
-    return self.env.user.company_id.id
-
-
-class mgmtsystem_system(models.Model):
-
-    _name = 'mgmtsystem.system'
-    _description = 'System'
-
-    name = fields.Char(string='System', required=True)
-    manual = fields.Many2one('document.page', string='Manual')
-    company_id = fields.Many2one('res.company', string='Company',
-                                 default=own_company)
+from . import mgmtsystem_system
