+++ mgmtsystem/views/board_mgmtsystem.xml
@@ -8,12 +8,7 @@
             <field name="model">board.board</field>
             <field name="arch" type="xml">
                 <form string="Management Systems Dashboard">
-                    <hpaned>
-                        <child1>
-                        </child1>
-                        <child2>
-                        </child2>
-                    </hpaned>
+                  
                 </form>
             </field>
         </record>
