+++ mgmtsystem/views/menus.xml
@@ -7,8 +7,7 @@
       name="Management System"
       sequence="150"
       groups="mgmtsystem.group_mgmtsystem_manager,base.group_user"
-      web_icon="images/mgmtsystem.png"
-      web_icon_hover="images/mgmtsystem-hover.png"
+
       />
 
     <menuitem
